move selectionvalue to Manager
