+++ src/es6/canvas_widget/KeySelectionValueAttribute.js
@@ -3,7 +3,7 @@ import "https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min.js"
 import _ from "lodash-es";
 import AbstractAttribute from "./AbstractAttribute";
 import Value from "./Value";
-import SelectionValue from "./SelectionValue";
+import { SelectionValue } from "./Manager";
 import loadHTML from "../html.template.loader";
 const keySelectionValueAttributeHtml = await loadHTML(
   "../../templates/canvas_widget/key_value_attribute.html",
